feat: rename GoTrueApi to GoTrueAdminApi
